moved SignalGroup from canmatrix to own package, with tests.
