+++ tests/test_canmatrix.py
@@ -36,73 +36,6 @@ def test_enum_defines_from_decimal():
 
 #
 
-# SignalGroup
-@pytest.fixture
-def the_group():
-    return canmatrix.canmatrix.SignalGroup(name="TestGroup", id=1)
-
-
-@pytest.fixture
-def some_signal():
-    return canmatrix.canmatrix.Signal(name="speed", size=8, factor=1.5)
-
-
-def test_signalgroup_empty(the_group):
-    assert [] == the_group.signals
-
-
-def test_signalgroup_can_add(the_group, some_signal):
-    the_group.add_signal(some_signal)
-    assert len(the_group.signals) == 1
-
-
-def test_signalgroup_can_remove(the_group, some_signal):
-    the_group.add_signal(some_signal)
-    the_group.del_signal(some_signal)
-    assert len(the_group.signals) == 0
-
-
-def test_signalgroup_no_duplicates(the_group, some_signal):
-    the_group.add_signal(some_signal)
-    the_group.add_signal(some_signal)
-    assert len(the_group.signals) == 1
-
-
-def test_signalgroup_is_iterable(the_group, some_signal):
-    the_group.add_signal(some_signal)
-    assert [some_signal] == [s for s in the_group]
-
-
-def test_signalgroup_find_something(the_group, some_signal):
-    the_group.add_signal(some_signal)
-    assert some_signal == the_group.by_name("speed")
-    assert some_signal == the_group["speed"]
-
-
-def test_signalgroup_find_nothing(the_group, some_signal):
-    the_group.add_signal(some_signal)
-    assert the_group.by_name("wrong") is None
-    with pytest.raises(KeyError):
-        _ = the_group["wrong"]
-
-
-def test_signalgroup_delete_nothing(the_group, some_signal):
-    the_group.add_signal(some_signal)
-    the_group.del_signal(canmatrix.canmatrix.Signal())
-    assert len(the_group.signals) == 1
-
-
-def test_encode_decode_frame():
-    input_data = {'signal': decimal.Decimal('3.5')}
-
-    s1 = canmatrix.canmatrix.Signal('signal', size=32, is_float=True)
-    f1 = canmatrix.canmatrix.Frame('frame', arbitration_id=1, size=4)
-    f1.add_signal(s1)
-
-    raw_bytes = f1.encode(input_data)
-    decoded_data = f1.decode(raw_bytes)
-
-    assert decoded_data['signal'].raw_value == float(input_data['signal'])
 
 
 # Frame tests
